GLib Reference

Public interface

Gtk4.GLib.bind_property — Function

bind_property(source::GObject, source_property, target::GObject, target_property, flags = BindingFlags_DEFAULT)

Creates a binding between source_property on source and target_property on target. When source_property is changed, target_property will be updated to the same value. Returns a GBinding object that can be used to release the binding using unbind_property.

See also unbind_property.

Related GTK function: g_object_bind_property

Gtk4.GLib.g_idle_add — Method

g_idle_add(f)

Add a Julia function f that will be called when there are no higher priority GTK events to be processed. This function can be used from any thread.

See also @idle_add.

Related GTK function: g_idle_add()

Gtk4.GLib.g_timeout_add — Method

g_timeout_add(f, interval)

Add a function f that will be called every interval milliseconds by the GTK main loop. The function is expected to return a Cint. If it returns 0, the function will not be called again. Otherwise it will be called the next time.

Related GTK function: g_timeout_add()

Gtk4.GLib.gtk_propertynames — Method

gtk_propertynames(w::GObject)

Prints a list of property names for the GObject w.

Gtk4.GLib.signal_handler_disconnect — Method

signal_handler_disconnect(w::GObject, id)

Disconnect a signal handler from a widget w by its id.

Gtk4.GLib.unbind_property — Method

unbind_property(b::GBinding)

Releases a binding created by bind_property.

See also bind_property.

Related GTK function: g_binding_unbind

Gtk4.GLib.@idle_add — Macro

@idle_add(ex)

Create a function from an expression ex that will be called when there are no higher priority GTK events to be processed. This function can be used from any thread.

See also g_idle_add.

Related GTK function: g_idle_add()

Private methods

Gtk4.GLib.g_type — Method

g_type(x)

Get the GType corresponding to a Julia type or object. See GLib documentation for more information.

Gtk4.GLib.propertyinfo — Method

propertyinfo(w::GObject, name)

Prints information about a property of the GObject w, including a brief description, its type, its default value, and its current value.

Gtk4.GLib.signal_handler_is_connected — Method

signal_handler_is_connected(widget, id) -> Bool

Return true/false depending on whether widget has a connected signal handler with the given id.

Gtk4.GLib.start_main_loop — Method

start_main_loop()

If the default GLib main event loop is not already running, start a Julia task that runs it.

See also stop_main_loop.

Gtk4.GLib.stop_main_loop — Method

stop_main_loop()

Stops the default GLib main loop after the next iteration. Does not affect loop operation if GApplication's run() method is being used instead of GLib.start_main_loop().

See also start_main_loop.