GLib Types

Gtk4.GLib.AppInfoCreateFlags — Type

See the GTK docs.

Gtk4.GLib.ApplicationFlags — Type

See the GTK docs.

Gtk4.GLib.AskPasswordFlags — Type

See the GTK docs.

Gtk4.GLib.BindingFlags — Type

See the GTK docs.

Gtk4.GLib.BookmarkFileError — Type

See the GTK docs.

Gtk4.GLib.BusNameOwnerFlags — Type

See the GTK docs.

Gtk4.GLib.BusNameWatcherFlags — Type

See the GTK docs.

Gtk4.GLib.BusType — Type

See the GTK docs.

Gtk4.GLib.ChecksumType — Type

See the GTK docs.

Gtk4.GLib.ConnectFlags — Type

See the GTK docs.

Gtk4.GLib.ConvertError — Type

See the GTK docs.

Gtk4.GLib.ConverterFlags — Type

See the GTK docs.

Gtk4.GLib.ConverterResult — Type

See the GTK docs.

Gtk4.GLib.CredentialsType — Type

See the GTK docs.

Gtk4.GLib.DBusCallFlags — Type

See the GTK docs.

Gtk4.GLib.DBusCapabilityFlags — Type

See the GTK docs.

Gtk4.GLib.DBusConnectionFlags — Type

See the GTK docs.

Gtk4.GLib.DBusError — Type

See the GTK docs.

Gtk4.GLib.DBusInterfaceSkeletonFlags — Type

See the GTK docs.

Gtk4.GLib.DBusMessageByteOrder — Type

See the GTK docs.

Gtk4.GLib.DBusMessageFlags — Type

See the GTK docs.

Gtk4.GLib.DBusMessageHeaderField — Type

See the GTK docs.

Gtk4.GLib.DBusMessageType — Type

See the GTK docs.

Gtk4.GLib.DBusObjectManagerClientFlags — Type

See the GTK docs.

Gtk4.GLib.DBusPropertyInfoFlags — Type

See the GTK docs.

Gtk4.GLib.DBusProxyFlags — Type

See the GTK docs.

Gtk4.GLib.DBusSendMessageFlags — Type

See the GTK docs.

Gtk4.GLib.DBusServerFlags — Type

See the GTK docs.

Gtk4.GLib.DBusSignalFlags — Type

See the GTK docs.

Gtk4.GLib.DBusSubtreeFlags — Type

See the GTK docs.

Gtk4.GLib.DataStreamByteOrder — Type

See the GTK docs.

Gtk4.GLib.DataStreamNewlineType — Type

See the GTK docs.

Gtk4.GLib.DateDMY — Type

See the GTK docs.

Gtk4.GLib.DateMonth — Type

See the GTK docs.

Gtk4.GLib.DateWeekday — Type

See the GTK docs.

Gtk4.GLib.DriveStartFlags — Type

See the GTK docs.

Gtk4.GLib.DriveStartStopType — Type

See the GTK docs.

Gtk4.GLib.EmblemOrigin — Type

See the GTK docs.

Gtk4.GLib.ErrorType — Type

See the GTK docs.

Gtk4.GLib.FileAttributeInfoFlags — Type

See the GTK docs.

Gtk4.GLib.FileAttributeStatus — Type

See the GTK docs.

Gtk4.GLib.FileAttributeType — Type

See the GTK docs.

Gtk4.GLib.FileCopyFlags — Type

See the GTK docs.

Gtk4.GLib.FileCreateFlags — Type

See the GTK docs.

Gtk4.GLib.FileError — Type

See the GTK docs.

Gtk4.GLib.FileMeasureFlags — Type

See the GTK docs.

Gtk4.GLib.FileMonitorEvent — Type

See the GTK docs.

Gtk4.GLib.FileMonitorFlags — Type

See the GTK docs.

Gtk4.GLib.FileQueryInfoFlags — Type

See the GTK docs.

Gtk4.GLib.FileSetContentsFlags — Type

See the GTK docs.

Gtk4.GLib.FileTest — Type

See the GTK docs.

Gtk4.GLib.FileType — Type

See the GTK docs.

Gtk4.GLib.FilesystemPreviewType — Type

See the GTK docs.

Gtk4.GLib.FormatSizeFlags — Type

See the GTK docs.

Gtk4.GLib.GAppInfoMonitor — Type

See the GTK docs.

Gtk4.GLib.GAppLaunchContext — Type

See the GTK docs.

Gtk4.GLib.GApplication — Type

See the GTK docs.

Gtk4.GLib.GApplicationCommandLine — Type

See the GTK docs.

Gtk4.GLib.GBinding — Type

See the GTK docs.

Gtk4.GLib.GBookmarkFile — Type

See the GTK docs.

Gtk4.GLib.GBufferedInputStream — Type

See the GTK docs.

Gtk4.GLib.GBufferedOutputStream — Type

See the GTK docs.

Gtk4.GLib.GBytes — Type

See the GTK docs.

Gtk4.GLib.GBytesIcon — Type

See the GTK docs.

Gtk4.GLib.GCancellable — Type

See the GTK docs.

Gtk4.GLib.GCharsetConverter — Type

See the GTK docs.

Gtk4.GLib.GChecksum — Type

See the GTK docs.

Gtk4.GLib.GConverterInputStream — Type

See the GTK docs.

Gtk4.GLib.GConverterOutputStream — Type

See the GTK docs.

Gtk4.GLib.GCredentials — Type

See the GTK docs.

Gtk4.GLib.GDBusActionGroup — Type

See the GTK docs.

Gtk4.GLib.GDBusAuthObserver — Type

See the GTK docs.

Gtk4.GLib.GDBusConnection — Type

See the GTK docs.

Gtk4.GLib.GDBusErrorEntry — Type

See the GTK docs.

Gtk4.GLib.GDBusInterfaceSkeleton — Type

See the GTK docs.

Gtk4.GLib.GDBusMenuModel — Type

See the GTK docs.

Gtk4.GLib.GDBusMessage — Type

See the GTK docs.

Gtk4.GLib.GDBusMethodInvocation — Type

See the GTK docs.

Gtk4.GLib.GDBusObjectManagerClient — Type

See the GTK docs.

Gtk4.GLib.GDBusObjectManagerServer — Type

See the GTK docs.

Gtk4.GLib.GDBusObjectProxy — Type

See the GTK docs.

Gtk4.GLib.GDBusObjectSkeleton — Type

See the GTK docs.

Gtk4.GLib.GDBusProxy — Type

See the GTK docs.

Gtk4.GLib.GDBusServer — Type

See the GTK docs.

Gtk4.GLib.GDataInputStream — Type

See the GTK docs.

Gtk4.GLib.GDataOutputStream — Type

See the GTK docs.

Gtk4.GLib.GDate — Type

See the GTK docs.

Gtk4.GLib.GDateTime — Type

See the GTK docs.

Gtk4.GLib.GDebugKey — Type

See the GTK docs.

Gtk4.GLib.GDesktopAppInfo — Type

See the GTK docs.

Gtk4.GLib.GEmblem — Type

See the GTK docs.

Gtk4.GLib.GEmblemedIcon — Type

See the GTK docs.

Gtk4.GLib.GEnumClass — Type

See the GTK docs.

Gtk4.GLib.GFileAttributeInfo — Type

See the GTK docs.

Gtk4.GLib.GFileAttributeInfoList — Type

See the GTK docs.

Gtk4.GLib.GFileAttributeMatcher — Type

See the GTK docs.

Gtk4.GLib.GFileEnumerator — Type

See the GTK docs.

Gtk4.GLib.GFileIOStream — Type

See the GTK docs.

Gtk4.GLib.GFileIcon — Type

See the GTK docs.

Gtk4.GLib.GFileInfo — Type

See the GTK docs.

Gtk4.GLib.GFileInputStream — Type

See the GTK docs.

Gtk4.GLib.GFileMonitor — Type

See the GTK docs.

Gtk4.GLib.GFileOutputStream — Type

See the GTK docs.

Gtk4.GLib.GFilenameCompleter — Type

See the GTK docs.

Gtk4.GLib.GFilterInputStream — Type

See the GTK docs.

Gtk4.GLib.GFilterOutputStream — Type

See the GTK docs.

Gtk4.GLib.GFlagsClass — Type

See the GTK docs.

Gtk4.GLib.GIOModule — Type

See the GTK docs.

Gtk4.GLib.GIOStream — Type

See the GTK docs.

Gtk4.GLib.GInetAddress — Type

See the GTK docs.

Gtk4.GLib.GInetAddressMask — Type

See the GTK docs.

Gtk4.GLib.GInetSocketAddress — Type

See the GTK docs.

Gtk4.GLib.GInitiallyUnowned — Type

See the GTK docs.

Gtk4.GLib.GInputStream — Type

See the GTK docs.

Gtk4.GLib.GInputVector — Type

See the GTK docs.

Gtk4.GLib.GInterfaceInfo — Type

See the GTK docs.

Gtk4.GLib.GKeyFile — Type

See the GTK docs.

Gtk4.GLib.GListStore — Type

See the GTK docs.

Gtk4.GLib.GListStore — Method

GListStore(t)

Create a GListStore, which is an array of GObjects that implements the GListModel interface. The argument t can be a Julia type corresponding to a GObject type (such as GtkWidget) or a symbol (such as :GtkWidget).

Gtk4.GLib.GLogField — Type

See the GTK docs.

Gtk4.GLib.GMainContext — Type

See the GTK docs.

Gtk4.GLib.GMainLoop — Type

See the GTK docs.

Gtk4.GLib.GMappedFile — Type

See the GTK docs.

Gtk4.GLib.GMarkupParseContext — Type

See the GTK docs.

Gtk4.GLib.GMatchInfo — Type

See the GTK docs.

Gtk4.GLib.GMemoryInputStream — Type

See the GTK docs.

Gtk4.GLib.GMemoryOutputStream — Type

See the GTK docs.

Gtk4.GLib.GMenu — Type

See the GTK docs.

Gtk4.GLib.GMenuAttributeIter — Type

See the GTK docs.

Gtk4.GLib.GMenuItem — Type

See the GTK docs.

Gtk4.GLib.GMenuLinkIter — Type

See the GTK docs.

Gtk4.GLib.GMenuModel — Type

See the GTK docs.

Gtk4.GLib.GMountOperation — Type

See the GTK docs.

Gtk4.GLib.GNativeSocketAddress — Type

See the GTK docs.

Gtk4.GLib.GNetworkAddress — Type

See the GTK docs.

Gtk4.GLib.GNetworkService — Type

See the GTK docs.

Gtk4.GLib.GNotification — Type

See the GTK docs.

Gtk4.GLib.GOptionEntry — Type

See the GTK docs.

Gtk4.GLib.GOutputStream — Type

See the GTK docs.

Gtk4.GLib.GOutputVector — Type

See the GTK docs.

Gtk4.GLib.GParamSpecTypeInfo — Type

See the GTK docs.

Gtk4.GLib.GPatternSpec — Type

See the GTK docs.

Gtk4.GLib.GPermission — Type

See the GTK docs.

Gtk4.GLib.GPropertyAction — Type

See the GTK docs.

Gtk4.GLib.GProxyAddress — Type

See the GTK docs.

Gtk4.GLib.GProxyAddressEnumerator — Type

See the GTK docs.

Gtk4.GLib.GRegex — Type

See the GTK docs.

Gtk4.GLib.GResolver — Type

See the GTK docs.

Gtk4.GLib.GResource — Type

See the GTK docs.

Gtk4.GLib.GScannerConfig — Type

See the GTK docs.

Gtk4.GLib.GSettings — Type

See the GTK docs.

Gtk4.GLib.GSettingsBackend — Type

See the GTK docs.

Gtk4.GLib.GSettingsSchema — Type

See the GTK docs.

Gtk4.GLib.GSettingsSchemaKey — Type

See the GTK docs.

Gtk4.GLib.GSettingsSchemaSource — Type

See the GTK docs.

Gtk4.GLib.GSignalInvocationHint — Type

See the GTK docs.

Gtk4.GLib.GSignalQuery — Type

See the GTK docs.

Gtk4.GLib.GSimpleAction — Method

GSimpleAction(name::AbstractString, 
              [parameter_type::Type{T}, [initial_state]]; kwargs...) where T

Create an action with a name and optionally a parameter_type from a Julia type (only a few simple types are supported) and an initial_state. If initial_state is not provided, the action will be stateless.

Keyword arguments set the action's GObject properties.

Gtk4.GLib.GSimpleActionGroup — Type

See the GTK docs.

Gtk4.GLib.GSimpleAsyncResult — Type

See the GTK docs.

Gtk4.GLib.GSimpleIOStream — Type

See the GTK docs.

Gtk4.GLib.GSimplePermission — Type

See the GTK docs.

Gtk4.GLib.GSimpleProxyResolver — Type

See the GTK docs.

Gtk4.GLib.GSocket — Type

See the GTK docs.

Gtk4.GLib.GSocketAddress — Type

See the GTK docs.

Gtk4.GLib.GSocketAddressEnumerator — Type

See the GTK docs.

Gtk4.GLib.GSocketClient — Type

See the GTK docs.

Gtk4.GLib.GSocketConnection — Type

See the GTK docs.

Gtk4.GLib.GSocketControlMessage — Type

See the GTK docs.

Gtk4.GLib.GSocketListener — Type

See the GTK docs.

Gtk4.GLib.GSocketService — Type

See the GTK docs.

Gtk4.GLib.GSource — Type

See the GTK docs.

Gtk4.GLib.GSourceCallbackFuncs — Type

See the GTK docs.

Gtk4.GLib.GSrvTarget — Type

See the GTK docs.

Gtk4.GLib.GString — Type

See the GTK docs.

Gtk4.GLib.GSubprocess — Type

See the GTK docs.

Gtk4.GLib.GSubprocessLauncher — Type

See the GTK docs.

Gtk4.GLib.GTask — Type

See the GTK docs.

Gtk4.GLib.GTcpConnection — Type

See the GTK docs.

Gtk4.GLib.GTcpWrapperConnection — Type

See the GTK docs.

Gtk4.GLib.GTestDBus — Type

See the GTK docs.

Gtk4.GLib.GThemedIcon — Type

See the GTK docs.

Gtk4.GLib.GThreadedResolver — Type

See the GTK docs.

Gtk4.GLib.GThreadedSocketService — Type

See the GTK docs.

Gtk4.GLib.GTimeZone — Type

See the GTK docs.

Gtk4.GLib.GTlsCertificate — Type

See the GTK docs.

Gtk4.GLib.GTlsConnection — Type

See the GTK docs.

Gtk4.GLib.GTlsDatabase — Type

See the GTK docs.

Gtk4.GLib.GTlsInteraction — Type

See the GTK docs.

Gtk4.GLib.GTlsPassword — Type

See the GTK docs.

Gtk4.GLib.GTypeFundamentalInfo — Type

See the GTK docs.

Gtk4.GLib.GTypeInfo — Type

See the GTK docs.

Gtk4.GLib.GTypeModule — Type

See the GTK docs.

Gtk4.GLib.GTypePluginClass — Type

See the GTK docs.

Gtk4.GLib.GTypeQuery — Type

See the GTK docs.

Gtk4.GLib.GUnixConnection — Type

See the GTK docs.

Gtk4.GLib.GUnixCredentialsMessage — Type

See the GTK docs.

Gtk4.GLib.GUri — Type

See the GTK docs.

Gtk4.GLib.GValueArray — Type

See the GTK docs.

Gtk4.GLib.GVariantBuilder — Type

See the GTK docs.

Gtk4.GLib.GVariantDict — Type

See the GTK docs.

Gtk4.GLib.GVariantType — Type

See the GTK docs.

Gtk4.GLib.GVfs — Type

See the GTK docs.

Gtk4.GLib.GVolumeMonitor — Type

See the GTK docs.

Gtk4.GLib.GXdpDocumentsIface — Type

See the GTK docs.

Gtk4.GLib.GXdpOpenURIIface — Type

See the GTK docs.

Gtk4.GLib.GXdpProxyResolverIface — Type

See the GTK docs.

Gtk4.GLib.GXdpTrashIface — Type

See the GTK docs.

Gtk4.GLib.GZlibCompressor — Type

See the GTK docs.

Gtk4.GLib.GZlibDecompressor — Type

See the GTK docs.

Gtk4.GLib.G_FreedesktopDBusIface — Type

See the GTK docs.

Gtk4.GLib.G_FreedesktopDBusProxy — Type

See the GTK docs.

Gtk4.GLib.G_FreedesktopDBusSkeleton — Type

See the GTK docs.

Gtk4.GLib.HookFlagMask — Type

See the GTK docs.

Gtk4.GLib.IOChannelError — Type

See the GTK docs.

Gtk4.GLib.IOCondition — Type

See the GTK docs.

Gtk4.GLib.IOError — Type

See the GTK docs.

Gtk4.GLib.IOErrorEnum — Type

See the GTK docs.

Gtk4.GLib.IOFlags — Type

See the GTK docs.

Gtk4.GLib.IOModuleScopeFlags — Type

See the GTK docs.

Gtk4.GLib.IOStatus — Type

See the GTK docs.

Gtk4.GLib.IOStreamSpliceFlags — Type

See the GTK docs.

Gtk4.GLib.KeyFileError — Type

See the GTK docs.

Gtk4.GLib.KeyFileFlags — Type

See the GTK docs.

Gtk4.GLib.LogLevelFlags — Type

See the GTK docs.

Gtk4.GLib.LogWriterOutput — Type

See the GTK docs.

Gtk4.GLib.MainContextFlags — Type

See the GTK docs.

Gtk4.GLib.MarkupCollectType — Type

See the GTK docs.

Gtk4.GLib.MarkupError — Type

See the GTK docs.

Gtk4.GLib.MarkupParseFlags — Type

See the GTK docs.

Gtk4.GLib.MemoryMonitorWarningLevel — Type

See the GTK docs.

Gtk4.GLib.MountMountFlags — Type

See the GTK docs.

Gtk4.GLib.MountOperationResult — Type

See the GTK docs.

Gtk4.GLib.MountUnmountFlags — Type

See the GTK docs.

Gtk4.GLib.NetworkConnectivity — Type

See the GTK docs.

Gtk4.GLib.NormalizeMode — Type

See the GTK docs.

Gtk4.GLib.NotificationPriority — Type

See the GTK docs.

Gtk4.GLib.NumberParserError — Type

See the GTK docs.

Gtk4.GLib.OnceStatus — Type

See the GTK docs.

Gtk4.GLib.OptionArg — Type

See the GTK docs.

Gtk4.GLib.OptionError — Type

See the GTK docs.

Gtk4.GLib.OptionFlags — Type

See the GTK docs.

Gtk4.GLib.OutputStreamSpliceFlags — Type

See the GTK docs.

Gtk4.GLib.ParamFlags — Type

See the GTK docs.

Gtk4.GLib.PasswordSave — Type

See the GTK docs.

Gtk4.GLib.PollableReturn — Type

See the GTK docs.

Gtk4.GLib.RegexCompileFlags — Type

See the GTK docs.

Gtk4.GLib.RegexError — Type

See the GTK docs.

Gtk4.GLib.RegexMatchFlags — Type

See the GTK docs.

Gtk4.GLib.ResolverError — Type

See the GTK docs.

Gtk4.GLib.ResolverNameLookupFlags — Type

See the GTK docs.

Gtk4.GLib.ResolverRecordType — Type

See the GTK docs.

Gtk4.GLib.ResourceError — Type

See the GTK docs.

Gtk4.GLib.ResourceFlags — Type

See the GTK docs.

Gtk4.GLib.ResourceLookupFlags — Type

See the GTK docs.

Gtk4.GLib.SeekType — Type

See the GTK docs.

Gtk4.GLib.SettingsBindFlags — Type

See the GTK docs.

Gtk4.GLib.ShellError — Type

See the GTK docs.

Gtk4.GLib.SignalFlags — Type

See the GTK docs.

Gtk4.GLib.SignalMatchType — Type

See the GTK docs.

Gtk4.GLib.SocketClientEvent — Type

See the GTK docs.

Gtk4.GLib.SocketFamily — Type

See the GTK docs.

Gtk4.GLib.SocketListenerEvent — Type

See the GTK docs.

Gtk4.GLib.SocketMsgFlags — Type

See the GTK docs.

Gtk4.GLib.SocketProtocol — Type

See the GTK docs.

Gtk4.GLib.SocketType — Type

See the GTK docs.

Gtk4.GLib.SpawnError — Type

See the GTK docs.

Gtk4.GLib.SpawnFlags — Type

See the GTK docs.

Gtk4.GLib.SubprocessFlags — Type

See the GTK docs.

Gtk4.GLib.TestDBusFlags — Type

See the GTK docs.

Gtk4.GLib.TestFileType — Type

See the GTK docs.

Gtk4.GLib.TestSubprocessFlags — Type

See the GTK docs.

Gtk4.GLib.ThreadError — Type

See the GTK docs.

Gtk4.GLib.TimeType — Type

See the GTK docs.

Gtk4.GLib.TlsAuthenticationMode — Type

See the GTK docs.

Gtk4.GLib.TlsCertificateFlags — Type

See the GTK docs.

Gtk4.GLib.TlsCertificateRequestFlags — Type

See the GTK docs.

Gtk4.GLib.TlsChannelBindingError — Type

See the GTK docs.

Gtk4.GLib.TlsChannelBindingType — Type

See the GTK docs.

Gtk4.GLib.TlsDatabaseLookupFlags — Type

See the GTK docs.

Gtk4.GLib.TlsDatabaseVerifyFlags — Type

See the GTK docs.

Gtk4.GLib.TlsError — Type

See the GTK docs.

Gtk4.GLib.TlsInteractionResult — Type

See the GTK docs.

Gtk4.GLib.TlsPasswordFlags — Type

See the GTK docs.

Gtk4.GLib.TlsProtocolVersion — Type

See the GTK docs.

Gtk4.GLib.TokenType — Type

See the GTK docs.

Gtk4.GLib.TraverseFlags — Type

See the GTK docs.

Gtk4.GLib.TraverseType — Type

See the GTK docs.

Gtk4.GLib.TypeFlags — Type

See the GTK docs.

Gtk4.GLib.TypeFundamentalFlags — Type

See the GTK docs.

Gtk4.GLib.UnicodeBreakType — Type

See the GTK docs.

Gtk4.GLib.UnicodeScript — Type

See the GTK docs.

Gtk4.GLib.UnicodeType — Type

See the GTK docs.

Gtk4.GLib.UnixSocketAddressType — Type

See the GTK docs.

Gtk4.GLib.UriError — Type

See the GTK docs.

Gtk4.GLib.UriFlags — Type

See the GTK docs.

Gtk4.GLib.UriHideFlags — Type

See the GTK docs.

Gtk4.GLib.UriParamsFlags — Type

See the GTK docs.

Gtk4.GLib.UserDirectory — Type

See the GTK docs.

Gtk4.GLib.VariantClass — Type

See the GTK docs.

Gtk4.GLib.VariantParseError — Type

See the GTK docs.

Gtk4.GLib.ZlibCompressorFormat — Type

See the GTK docs.