Gtk4 Types

Types

Gtk4.AccessibleAnnouncementPriority — Type

See the GTK docs.

Gtk4.AccessibleAutocomplete — Type

See the GTK docs.

Gtk4.AccessibleInvalidState — Type

See the GTK docs.

Gtk4.AccessiblePlatformState — Type

See the GTK docs.

Gtk4.AccessibleProperty — Type

See the GTK docs.

Gtk4.AccessibleRelation — Type

See the GTK docs.

Gtk4.AccessibleRole — Type

See the GTK docs.

Gtk4.AccessibleSort — Type

See the GTK docs.

Gtk4.AccessibleState — Type

See the GTK docs.

Gtk4.AccessibleTextContentChange — Type

See the GTK docs.

Gtk4.AccessibleTextGranularity — Type

See the GTK docs.

Gtk4.AccessibleTristate — Type

See the GTK docs.

Gtk4.Align — Type

See the GTK docs.

Gtk4.AnchorHints — Type

See the GTK docs.

Gtk4.ApplicationInhibitFlags — Type

See the GTK docs.

Gtk4.ArrowType — Type

See the GTK docs.

Gtk4.AssistantPageType — Type

See the GTK docs.

Gtk4.AxisFlags — Type

See the GTK docs.

Gtk4.AxisUse — Type

See the GTK docs.

Gtk4.BaselinePosition — Type

See the GTK docs.

Gtk4.BorderStyle — Type

See the GTK docs.

Gtk4.BuilderClosureFlags — Type

See the GTK docs.

Gtk4.BuilderError — Type

See the GTK docs.

Gtk4.ButtonsType — Type

See the GTK docs.

Gtk4.CellRendererAccelMode — Type

See the GTK docs.

Gtk4.CellRendererMode — Type

See the GTK docs.

Gtk4.CellRendererState — Type

See the GTK docs.

Gtk4.Collation — Type

See the GTK docs.

Gtk4.ConstraintAttribute — Type

See the GTK docs.

Gtk4.ConstraintRelation — Type

See the GTK docs.

Gtk4.ConstraintStrength — Type

See the GTK docs.

Gtk4.ConstraintVflParserError — Type

See the GTK docs.

Gtk4.ContentFit — Type

See the GTK docs.

Gtk4.CornerType — Type

See the GTK docs.

Gtk4.CrossingMode — Type

See the GTK docs.

Gtk4.CssParserError — Type

See the GTK docs.

Gtk4.CssParserWarning — Type

See the GTK docs.

Gtk4.DebugFlags — Type

See the GTK docs.

Gtk4.DeleteType — Type

See the GTK docs.

Gtk4.DevicePadFeature — Type

See the GTK docs.

Gtk4.DeviceToolType — Type

See the GTK docs.

Gtk4.DialogError — Type

See the GTK docs.

Gtk4.DialogFlags — Type

See the GTK docs.

Gtk4.DirectionType — Type

See the GTK docs.

Gtk4.DmabufError — Type

See the GTK docs.

Gtk4.DragAction — Type

See the GTK docs.

Gtk4.DragCancelReason — Type

See the GTK docs.

Gtk4.EditableProperties — Type

See the GTK docs.

Gtk4.EntryIconPosition — Type

See the GTK docs.

Gtk4.EventControllerScrollFlags — Type

See the GTK docs.

Gtk4.EventSequenceState — Type

See the GTK docs.

Gtk4.EventType — Type

See the GTK docs.

Gtk4.FileChooserAction — Type

See the GTK docs.

Gtk4.FileChooserError — Type

See the GTK docs.

Gtk4.FilterChange — Type

See the GTK docs.

Gtk4.FilterMatch — Type

See the GTK docs.

Gtk4.FontChooserLevel — Type

See the GTK docs.

Gtk4.FontLevel — Type

See the GTK docs.

Gtk4.FrameClockPhase — Type

See the GTK docs.

Gtk4.FullscreenMode — Type

See the GTK docs.

Gtk4.GLAPI — Type

See the GTK docs.

Gtk4.GLError — Type

See the GTK docs.

Gtk4.GdkAppLaunchContext — Type

See the GTK docs.

Gtk4.GdkButtonEvent — Type

See the GTK docs.

Gtk4.GdkCairoContext — Type

See the GTK docs.

Gtk4.GdkClipboard — Type

See the GTK docs.

Gtk4.GdkContentDeserializer — Type

See the GTK docs.

Gtk4.GdkContentFormats — Type

See the GTK docs.

Gtk4.GdkContentFormatsBuilder — Type

See the GTK docs.

Gtk4.GdkContentProvider — Type

See the GTK docs.

Gtk4.GdkContentSerializer — Type

See the GTK docs.

Gtk4.GdkCrossingEvent — Type

See the GTK docs.

Gtk4.GdkCursor — Type

See the GTK docs.

Gtk4.GdkDNDEvent — Type

See the GTK docs.

Gtk4.GdkDeleteEvent — Type

See the GTK docs.

Gtk4.GdkDevice — Type

See the GTK docs.

Gtk4.GdkDeviceTool — Type

See the GTK docs.

Gtk4.GdkDisplay — Type

See the GTK docs.

Gtk4.GdkDisplay — Method

GdkDisplay()

Get the default GdkDisplay.

Related GDK function: gdk_display_get_default()

Gtk4.GdkDisplayManager — Type

See the GTK docs.

Gtk4.GdkDmabufFormats — Type

See the GTK docs.

Gtk4.GdkDmabufTexture — Type

See the GTK docs.

Gtk4.GdkDmabufTextureBuilder — Type

See the GTK docs.

Gtk4.GdkDrag — Type

See the GTK docs.

Gtk4.GdkDragSurfaceSize — Type

See the GTK docs.

Gtk4.GdkDrawContext — Type

See the GTK docs.

Gtk4.GdkDrop — Type

See the GTK docs.

Gtk4.GdkEvent — Type

See the GTK docs.

Gtk4.GdkEventSequence — Type

See the GTK docs.

Gtk4.GdkFileList — Type

See the GTK docs.

Gtk4.GdkFocusEvent — Type

See the GTK docs.

Gtk4.GdkFrameClock — Type

See the GTK docs.

Gtk4.GdkFrameTimings — Type

See the GTK docs.

Gtk4.GdkGLContext — Type

See the GTK docs.

Gtk4.GdkGLTexture — Type

See the GTK docs.

Gtk4.GdkGLTextureBuilder — Type

See the GTK docs.

Gtk4.GdkGrabBrokenEvent — Type

See the GTK docs.

Gtk4.GdkKeyEvent — Type

See the GTK docs.

Gtk4.GdkKeymapKey — Type

See the GTK docs.

Gtk4.GdkMemoryTexture — Type

See the GTK docs.

Gtk4.GdkMemoryTexture — Type

GdkMemoryTexture(img::Array, tp = true)

Creates a GdkMemoryTexture, copying an image array. If tp is set to true, the image will be transposed before copying so that the texture's orientation when displayed by GTK widgets like GtkPicture will match how the image is displayed in Julia apps like ImageShow.

Gtk4.GdkMonitor — Type

See the GTK docs.

Gtk4.GdkMotionEvent — Type

See the GTK docs.

Gtk4.GdkPadEvent — Type

See the GTK docs.

Gtk4.GdkPopupLayout — Type

See the GTK docs.

Gtk4.GdkProximityEvent — Type

See the GTK docs.

Gtk4.GdkRGBA — Type

See the GTK docs.

Gtk4.GdkRectangle — Type

See the GTK docs.

Gtk4.GdkScrollEvent — Type

See the GTK docs.

Gtk4.GdkSeat — Type

See the GTK docs.

Gtk4.GdkSnapshot — Type

See the GTK docs.

Gtk4.GdkSurface — Type

See the GTK docs.

Gtk4.GdkTexture — Type

See the GTK docs.

Gtk4.GdkTextureDownloader — Type

See the GTK docs.

Gtk4.GdkTimeCoord — Type

See the GTK docs.

Gtk4.GdkToplevelLayout — Type

See the GTK docs.

Gtk4.GdkTouchEvent — Type

See the GTK docs.

Gtk4.GdkTouchpadEvent — Type

See the GTK docs.

Gtk4.GdkVulkanContext — Type

See the GTK docs.

Gtk4.GraphicsOffloadEnabled — Type

See the GTK docs.

Gtk4.Gravity — Type

See the GTK docs.

Gtk4.GskBlendNode — Type

See the GTK docs.

Gtk4.GskBlurNode — Type

See the GTK docs.

Gtk4.GskBorderNode — Type

See the GTK docs.

Gtk4.GskCairoNode — Type

See the GTK docs.

Gtk4.GskCairoRenderer — Type

See the GTK docs.

Gtk4.GskClipNode — Type

See the GTK docs.

Gtk4.GskColorMatrixNode — Type

See the GTK docs.

Gtk4.GskColorNode — Type

See the GTK docs.

Gtk4.GskColorStop — Type

See the GTK docs.

Gtk4.GskConicGradientNode — Type

See the GTK docs.

Gtk4.GskContainerNode — Type

See the GTK docs.

Gtk4.GskCrossFadeNode — Type

See the GTK docs.

Gtk4.GskDebugNode — Type

See the GTK docs.

Gtk4.GskGLShader — Type

See the GTK docs.

Gtk4.GskGLShaderNode — Type

See the GTK docs.

Gtk4.GskInsetShadowNode — Type

See the GTK docs.

Gtk4.GskLinearGradientNode — Type

See the GTK docs.

Gtk4.GskMaskNode — Type

See the GTK docs.

Gtk4.GskOpacityNode — Type

See the GTK docs.

Gtk4.GskOutsetShadowNode — Type

See the GTK docs.

Gtk4.GskParseLocation — Type

See the GTK docs.

Gtk4.GskRadialGradientNode — Type

See the GTK docs.

Gtk4.GskRenderNode — Type

See the GTK docs.

Gtk4.GskRenderer — Type

See the GTK docs.

Gtk4.GskRepeatNode — Type

See the GTK docs.

Gtk4.GskRepeatingLinearGradientNode — Type

See the GTK docs.

Gtk4.GskRepeatingRadialGradientNode — Type

See the GTK docs.

Gtk4.GskRoundedClipNode — Type

See the GTK docs.

Gtk4.GskRoundedRect — Type

See the GTK docs.

Gtk4.GskShaderArgsBuilder — Type

See the GTK docs.

Gtk4.GskShadow — Type

See the GTK docs.

Gtk4.GskShadowNode — Type

See the GTK docs.

Gtk4.GskTextNode — Type

See the GTK docs.

Gtk4.GskTextureNode — Type

See the GTK docs.

Gtk4.GskTextureScaleNode — Type

See the GTK docs.

Gtk4.GskTransform — Type

See the GTK docs.

Gtk4.GskTransformNode — Type

See the GTK docs.

Gtk4.GskVulkanRenderer — Type

See the GTK docs.

Gtk4.GtkATContext — Type

See the GTK docs.

Gtk4.GtkAboutDialog — Type

See the GTK docs.

Gtk4.GtkAccessibleList — Type

See the GTK docs.

Gtk4.GtkAccessibleTextRange — Type

See the GTK docs.

Gtk4.GtkActionBar — Type

See the GTK docs.

Gtk4.GtkActivateAction — Type

See the GTK docs.

Gtk4.GtkAdjustment — Type

See the GTK docs.

Gtk4.GtkAlertDialog — Type

See the GTK docs.

Gtk4.GtkAlternativeTrigger — Type

See the GTK docs.

Gtk4.GtkAnyFilter — Type

See the GTK docs.

Gtk4.GtkAppChooserButton — Type

See the GTK docs.

Gtk4.GtkAppChooserDialog — Type

See the GTK docs.

Gtk4.GtkAppChooserWidget — Type

See the GTK docs.

Gtk4.GtkApplication — Type

GtkApplication(id = nothing, flags = GLib.ApplicationFlags_FLAGS_NONE)

Create a GtkApplication with DBus id id and flags.

Related GTK function: gtk_application_new()

Gtk4.GtkApplication — Type

See the GTK docs.

Gtk4.GtkApplicationWindow — Type

See the GTK docs.

Gtk4.GtkApplicationWindow — Method

GtkApplicationWindow(app::GtkApplication, title::AbstractString; kwargs...)

Create an empty GtkApplicationWindow for a GtkApplication app and a title. Keyword arguments can be used to set GObject properties.

GTK docs: GtkApplicationWindow

Gtk4.GtkAspectFrame — Type

See the GTK docs.

Gtk4.GtkAssistant — Type

See the GTK docs.

Gtk4.GtkAssistantPage — Type

See the GTK docs.

Gtk4.GtkBinLayout — Type

See the GTK docs.

Gtk4.GtkBitset — Type

See the GTK docs.

Gtk4.GtkBitsetIter — Type

See the GTK docs.

Gtk4.GtkBookmarkList — Type

See the GTK docs.

Gtk4.GtkBoolFilter — Type

See the GTK docs.

Gtk4.GtkBorder — Type

See the GTK docs.

Gtk4.GtkBox — Type

GtkBox(orientation::Symbol, spacing::Integer=0; kwargs...)

Create and return a GtkBox widget. The orientation argument can be :h for horizontal, or :v for vertical. The spacing argument controls the spacing between child widgets in pixels. Keyword arguments allow you to set GObject properties.

Gtk4.GtkBox — Type

See the GTK docs.

Gtk4.GtkBoxLayout — Type

See the GTK docs.

Gtk4.GtkBuildableParseContext — Type

See the GTK docs.

Gtk4.GtkBuildableParser — Type

See the GTK docs.

Gtk4.GtkBuilder — Type

GtkBuilder(; kwargs...)
GtkBuilder(filename::AbstractString; kwargs...)
GtkBuilder(string::AbstractString, _length::Integer; kwargs...)

Create a GtkBuilder object. If filename is given (the constructor with a single string argument), XML describing the user interface will be read from a file. If string and length are given (the constructor with a string and an integer), XML will be read from a string of a certain length. If length is -1 the full string will be used.

See the GTK docs.

Gtk4.GtkBuilderCScope — Type

See the GTK docs.

Gtk4.GtkBuilderListItemFactory — Type

See the GTK docs.

Gtk4.GtkButton — Type

See the GTK docs.

Gtk4.GtkButton — Method

GtkButton(w::GtkWidget)

Create a GtkButton and add a widget w as its child.

Gtk4.GtkButton — Method

GtkButton(s::Symbol, str::AbstractString)

Create and return a GtkButton widget.

If s is :label, create a button with a string label.

If s is :mnemonic, create a button with a string label, where the first letter preceded by an underscore character defines a mnemonic. Pressing Alt and that letter activates the button.

If s is :icon_name, create a button with an icon from the current icon theme.

Related GTK functions: gtk_button_new_with_label(), gtk_button_new_with_mnemonic(), gtk_button_new_from_icon_name()

Gtk4.GtkCalendar — Type

See the GTK docs.

Gtk4.GtkCallbackAction — Type

See the GTK docs.

Gtk4.GtkCellArea — Type

See the GTK docs.

Gtk4.GtkCellAreaBox — Type

See the GTK docs.

Gtk4.GtkCellAreaContext — Type

See the GTK docs.

Gtk4.GtkCellRenderer — Type

See the GTK docs.

Gtk4.GtkCellRendererAccel — Type

See the GTK docs.

Gtk4.GtkCellRendererCombo — Type

See the GTK docs.

Gtk4.GtkCellRendererPixbuf — Type

See the GTK docs.

Gtk4.GtkCellRendererProgress — Type

See the GTK docs.

Gtk4.GtkCellRendererSpin — Type

See the GTK docs.

Gtk4.GtkCellRendererSpinner — Type

See the GTK docs.

Gtk4.GtkCellRendererText — Type

See the GTK docs.

Gtk4.GtkCellRendererToggle — Type

See the GTK docs.

Gtk4.GtkCellView — Type

See the GTK docs.

Gtk4.GtkCenterBox — Type

See the GTK docs.

Gtk4.GtkCenterBox — Method

GtkCenterBox(orientation::Symbol; kwargs...)

Create and return a GtkCenterBox widget. The orientation argument can be :h for horizontal, or :v for vertical. Keyword arguments allow you to set GObject properties.

Gtk4.GtkCenterLayout — Type

See the GTK docs.

Gtk4.GtkCheckButton — Type

See the GTK docs.

Gtk4.GtkColorButton — Type

See the GTK docs.

Gtk4.GtkColorChooserDialog — Type

See the GTK docs.

Gtk4.GtkColorChooserWidget — Type

See the GTK docs.

Gtk4.GtkColorDialog — Type

See the GTK docs.

Gtk4.GtkColorDialogButton — Type

See the GTK docs.

Gtk4.GtkColumnView — Type

See the GTK docs.

Gtk4.GtkColumnViewCell — Type

See the GTK docs.

Gtk4.GtkColumnViewColumn — Type

See the GTK docs.

Gtk4.GtkColumnViewRow — Type

See the GTK docs.

Gtk4.GtkColumnViewSorter — Type

See the GTK docs.

Gtk4.GtkComboBox — Type

See the GTK docs.

Gtk4.GtkComboBoxText — Type

See the GTK docs.

Gtk4.GtkConstantExpression — Type

See the GTK docs.

Gtk4.GtkConstraint — Type

See the GTK docs.

Gtk4.GtkConstraintGuide — Type

See the GTK docs.

Gtk4.GtkConstraintLayout — Type

See the GTK docs.

Gtk4.GtkConstraintLayoutChild — Type

See the GTK docs.

Gtk4.GtkCssProvider — Type

GtkCssProvider(data, filename = nothing)

Create a GtkCssProvider object using CSS from a string data. If data is set to nothing, CSS is instead loaded from a file filename. If both arguments are nothing, an empty GtkCssProvider is returned.

Related GTK functions: gtk_css_provider_load_from_path(), gtk_css_provider_load_from_data()

Gtk4.GtkCssProvider — Type

See the GTK docs.

Gtk4.GtkCssSection — Type

See the GTK docs.

Gtk4.GtkCustomFilter — Type

See the GTK docs.

Gtk4.GtkCustomLayout — Type

See the GTK docs.

Gtk4.GtkCustomSorter — Type

See the GTK docs.

Gtk4.GtkDialog — Type

See the GTK docs.

Gtk4.GtkDirectoryList — Type

See the GTK docs.

Gtk4.GtkDragIcon — Type

See the GTK docs.

Gtk4.GtkDragSource — Type

See the GTK docs.

Gtk4.GtkDrawingArea — Type

See the GTK docs.

Gtk4.GtkDropControllerMotion — Type

See the GTK docs.

Gtk4.GtkDropDown — Type

See the GTK docs.

Gtk4.GtkDropDown — Method

GtkDropDown(a::AbstractArray; kwargs...)

Create a dropdown widget with a GtkStringList as its model. The model will be populated with the elements of a converted to strings. Keyword arguments set GObject properties.

Gtk4.GtkDropDown — Method

GtkDropDown(; kwargs...)

Create a dropdown widget with no model (and thus no options to selected). A model can be added using model. Keyword arguments set GObject properties.

Gtk4.GtkDropTarget — Type

See the GTK docs.

Gtk4.GtkDropTargetAsync — Type

See the GTK docs.

Gtk4.GtkEditableLabel — Type

See the GTK docs.

Gtk4.GtkEmojiChooser — Type

See the GTK docs.

Gtk4.GtkEntry — Type

See the GTK docs.

Gtk4.GtkEntryBuffer — Type

See the GTK docs.

Gtk4.GtkEntryCompletion — Type

See the GTK docs.

Gtk4.GtkEventController — Type

See the GTK docs.

Gtk4.GtkEventControllerFocus — Type

See the GTK docs.

Gtk4.GtkEventControllerKey — Type

See the GTK docs.

Gtk4.GtkEventControllerLegacy — Type

See the GTK docs.

Gtk4.GtkEventControllerMotion — Type

See the GTK docs.

Gtk4.GtkEventControllerScroll — Type

See the GTK docs.

Gtk4.GtkEveryFilter — Type

See the GTK docs.

Gtk4.GtkExpander — Type

See the GTK docs.

Gtk4.GtkExpression — Type

See the GTK docs.

Gtk4.GtkExpressionWatch — Type

See the GTK docs.

Gtk4.GtkFileChooserDialog — Type

See the GTK docs.

Gtk4.GtkFileChooserNative — Type

See the GTK docs.

Gtk4.GtkFileChooserWidget — Type

See the GTK docs.

Gtk4.GtkFileDialog — Type

See the GTK docs.

Gtk4.GtkFileDialog — Method

GtkFileDialog(; kwargs...)

Selected keyword arguments

accept_label: the text to show on the dialog's accept button

default_filter: the GtkFileFilter initially active in the file dialog

filters: a GListModel of file filters

initial_name: the filename or directory that is initially selected in the file chooser dialog

title: the title of the dialog

modal: whether the dialog is modal

Gtk4.GtkFileFilter — Type

See the GTK docs.

Gtk4.GtkFileLauncher — Type

See the GTK docs.

Gtk4.GtkFilter — Type

See the GTK docs.

Gtk4.GtkFilterListModel — Type

See the GTK docs.

Gtk4.GtkFixed — Type

See the GTK docs.

Gtk4.GtkFixedLayout — Type

See the GTK docs.

Gtk4.GtkFixedLayoutChild — Type

See the GTK docs.

Gtk4.GtkFlattenListModel — Type

See the GTK docs.

Gtk4.GtkFlowBox — Type

See the GTK docs.

Gtk4.GtkFlowBoxChild — Type

See the GTK docs.

Gtk4.GtkFontButton — Type

See the GTK docs.

Gtk4.GtkFontChooserDialog — Type

See the GTK docs.

Gtk4.GtkFontChooserWidget — Type

See the GTK docs.

Gtk4.GtkFontDialog — Type

See the GTK docs.

Gtk4.GtkFontDialogButton — Type

See the GTK docs.

Gtk4.GtkFrame — Type

GtkFrame(w::GtkWidget, label=nothing; kwargs...)

Create a GtkFrame with an optional string label and add w as its child. Keyword arguments allow you to set GObject properties.

Gtk4.GtkFrame — Type

See the GTK docs.

Gtk4.GtkFrame — Method

GtkFrame(label=nothing; kwargs...)

Create a GtkFrame, a layout widget that can hold a single child widget, with an optional string label. Keyword arguments allow you to set GObject properties.

Gtk4.GtkGLArea — Type

See the GTK docs.

Gtk4.GtkGesture — Type

See the GTK docs.

Gtk4.GtkGestureClick — Type

See the GTK docs.

Gtk4.GtkGestureDrag — Type

See the GTK docs.

Gtk4.GtkGestureLongPress — Type

See the GTK docs.

Gtk4.GtkGesturePan — Type

See the GTK docs.

Gtk4.GtkGestureRotate — Type

See the GTK docs.

Gtk4.GtkGestureSingle — Type

See the GTK docs.

Gtk4.GtkGestureStylus — Type

See the GTK docs.

Gtk4.GtkGestureSwipe — Type

See the GTK docs.

Gtk4.GtkGestureZoom — Type

See the GTK docs.

Gtk4.GtkGraphicsOffload — Type

See the GTK docs.

Gtk4.GtkGrid — Type

See the GTK docs.

Gtk4.GtkGridLayout — Type

See the GTK docs.

Gtk4.GtkGridLayoutChild — Type

See the GTK docs.

Gtk4.GtkGridView — Type

GtkGridView(model=nothing; kwargs...)

Create a GtkGridView widget, optionally with a model. Keyword arguments set GObject properties.

Gtk4.GtkGridView — Type

See the GTK docs.

Gtk4.GtkHeaderBar — Type

See the GTK docs.

Gtk4.GtkIMContext — Type

See the GTK docs.

Gtk4.GtkIMContextSimple — Type

See the GTK docs.

Gtk4.GtkIMMulticontext — Type

See the GTK docs.

Gtk4.GtkIconPaintable — Type

See the GTK docs.

Gtk4.GtkIconTheme — Type

See the GTK docs.

Gtk4.GtkIconTheme — Method

GtkIconTheme(d::GdkDisplay)

Get the icon theme for a GdkDisplay.

Related GTK function: gtk_icon_theme_get_for_display()

Gtk4.GtkIconView — Type

See the GTK docs.

Gtk4.GtkImage — Type

See the GTK docs.

Gtk4.GtkInfoBar — Type

See the GTK docs.

Gtk4.GtkInscription — Type

See the GTK docs.

Gtk4.GtkKeyvalTrigger — Type

See the GTK docs.

Gtk4.GtkLabel — Type

See the GTK docs.

Gtk4.GtkLayoutChild — Type

See the GTK docs.

Gtk4.GtkLayoutManager — Type

See the GTK docs.

Gtk4.GtkLevelBar — Type

See the GTK docs.

Gtk4.GtkLinkButton — Type

See the GTK docs.

Gtk4.GtkListBase — Type

See the GTK docs.

Gtk4.GtkListBox — Type

See the GTK docs.

Gtk4.GtkListBoxRow — Type

See the GTK docs.

Gtk4.GtkListHeader — Type

See the GTK docs.

Gtk4.GtkListItem — Type

See the GTK docs.

Gtk4.GtkListItemFactory — Type

See the GTK docs.

Gtk4.GtkListStore — Type

See the GTK docs.

Gtk4.GtkListView — Type

GtkListView(model=nothing; kwargs...)

Create a GtkListView widget, optionally with a model. Keyword arguments set GObject properties.

Gtk4.GtkListView — Type

See the GTK docs.

Gtk4.GtkLockButton — Type

See the GTK docs.

Gtk4.GtkMapListModel — Type

See the GTK docs.

Gtk4.GtkMediaControls — Type

See the GTK docs.

Gtk4.GtkMediaFile — Type

See the GTK docs.

Gtk4.GtkMediaStream — Type

See the GTK docs.

Gtk4.GtkMenuButton — Type

See the GTK docs.

Gtk4.GtkMessageDialog — Type

See the GTK docs.

Gtk4.GtkMnemonicAction — Type

See the GTK docs.

Gtk4.GtkMnemonicTrigger — Type

See the GTK docs.

Gtk4.GtkMountOperation — Type

See the GTK docs.

Gtk4.GtkMultiFilter — Type

See the GTK docs.

Gtk4.GtkMultiSelection — Type

See the GTK docs.

Gtk4.GtkMultiSorter — Type

See the GTK docs.

Gtk4.GtkNamedAction — Type

See the GTK docs.

Gtk4.GtkNativeDialog — Type

See the GTK docs.

Gtk4.GtkNeverTrigger — Type

See the GTK docs.

Gtk4.GtkNoSelection — Type

See the GTK docs.

Gtk4.GtkNotebook — Type

See the GTK docs.

Gtk4.GtkNotebookPage — Type

See the GTK docs.

Gtk4.GtkNothingAction — Type

See the GTK docs.

Gtk4.GtkNumericSorter — Type

See the GTK docs.

Gtk4.GtkObjectExpression — Type

See the GTK docs.

Gtk4.GtkOverlay — Type

See the GTK docs.

Gtk4.GtkOverlay — Method

GtkOverlay(w=nothing; kwargs...)

Create a GtkOverlay, a layout widget that holds one main child and other child "overlay" widgets that are drawn on top of the main child. The main child can be set using the argument w.

Gtk4.GtkOverlayLayout — Type

See the GTK docs.

Gtk4.GtkOverlayLayoutChild — Type

See the GTK docs.

Gtk4.GtkPadActionEntry — Type

See the GTK docs.

Gtk4.GtkPadController — Type

See the GTK docs.

Gtk4.GtkPageSetup — Type

See the GTK docs.

Gtk4.GtkPaned — Type

See the GTK docs.

Gtk4.GtkPaned — Method

GtkPaned(orientation::Symbol; kwargs...)

Create and return a GtkPaned widget. The orientation argument can be :h for horizontal, or :v for vertical. Keyword arguments allow you to set GObject properties.

Gtk4.GtkPaperSize — Type

See the GTK docs.

Gtk4.GtkPasswordEntry — Type

See the GTK docs.

Gtk4.GtkPasswordEntryBuffer — Type

See the GTK docs.

Gtk4.GtkPicture — Type

See the GTK docs.

Gtk4.GtkPopover — Type

See the GTK docs.

Gtk4.GtkPopoverMenu — Type

See the GTK docs.

Gtk4.GtkPopoverMenuBar — Type

See the GTK docs.

Gtk4.GtkPrintContext — Type

See the GTK docs.

Gtk4.GtkPrintDialog — Type

See the GTK docs.

Gtk4.GtkPrintJob — Type

See the GTK docs.

Gtk4.GtkPrintOperation — Type

See the GTK docs.

Gtk4.GtkPrintSettings — Type

See the GTK docs.

Gtk4.GtkPrintSetup — Type

See the GTK docs.

Gtk4.GtkPrinter — Type

See the GTK docs.

Gtk4.GtkProgressBar — Type

See the GTK docs.

Gtk4.GtkPropertyExpression — Type

See the GTK docs.

Gtk4.GtkRange — Type

See the GTK docs.

Gtk4.GtkRecentData — Type

See the GTK docs.

Gtk4.GtkRecentInfo — Type

See the GTK docs.

Gtk4.GtkRecentManager — Type

See the GTK docs.

Gtk4.GtkRequestedSize — Type

See the GTK docs.

Gtk4.GtkRequisition — Type

See the GTK docs.

Gtk4.GtkRevealer — Type

See the GTK docs.

Gtk4.GtkScale — Type

See the GTK docs.

Gtk4.GtkScale — Method

GtkScale(orientation, [scale::AbstractRange]; kwargs...)

Create a scale widget with horizontal (:h) or vertical (:v) orientation and an optional range. Keyword arguments can be used to set properties.

Gtk4.GtkScaleButton — Type

See the GTK docs.

Gtk4.GtkScrollInfo — Type

See the GTK docs.

Gtk4.GtkScrollbar — Type

See the GTK docs.

Gtk4.GtkScrolledWindow — Type

See the GTK docs.

Gtk4.GtkSearchBar — Type

See the GTK docs.

Gtk4.GtkSearchEntry — Type

See the GTK docs.

Gtk4.GtkSelectionFilterModel — Type

See the GTK docs.

Gtk4.GtkSeparator — Type

See the GTK docs.

Gtk4.GtkSettings — Type

See the GTK docs.

Gtk4.GtkShortcut — Type

See the GTK docs.

Gtk4.GtkShortcutAction — Type

See the GTK docs.

Gtk4.GtkShortcutController — Type

See the GTK docs.

Gtk4.GtkShortcutLabel — Type

See the GTK docs.

Gtk4.GtkShortcutTrigger — Type

See the GTK docs.

Gtk4.GtkShortcutsGroup — Type

See the GTK docs.

Gtk4.GtkShortcutsSection — Type

See the GTK docs.

Gtk4.GtkShortcutsShortcut — Type

See the GTK docs.

Gtk4.GtkShortcutsWindow — Type

See the GTK docs.

Gtk4.GtkSignalAction — Type

See the GTK docs.

Gtk4.GtkSignalListItemFactory — Type

See the GTK docs.

Gtk4.GtkSignalListItemFactory — Method

GtkSignalListItemFactory(setup_cb, bind_cb)

Create a GtkSignalListItemFactory and immediately connect "setup" and "bind" callback functions setup_cb and bind_cb, respectively.

Gtk4.GtkSingleSelection — Type

See the GTK docs.

Gtk4.GtkSizeGroup — Type

See the GTK docs.

Gtk4.GtkSliceListModel — Type

See the GTK docs.

Gtk4.GtkSnapshot — Type

See the GTK docs.

Gtk4.GtkSortListModel — Type

See the GTK docs.

Gtk4.GtkSorter — Type

See the GTK docs.

Gtk4.GtkSpinButton — Type

See the GTK docs.

Gtk4.GtkSpinner — Type

See the GTK docs.

Gtk4.GtkStack — Type

See the GTK docs.

Gtk4.GtkStackPage — Type

See the GTK docs.

Gtk4.GtkStackSidebar — Type

See the GTK docs.

Gtk4.GtkStackSwitcher — Type

See the GTK docs.

Gtk4.GtkStatusbar — Type

See the GTK docs.

Gtk4.GtkStringFilter — Type

See the GTK docs.

Gtk4.GtkStringList — Type

See the GTK docs.

Gtk4.GtkStringList — Method

GtkStringList()

Create an empty GtkStringList, which implements the GListModel interface and holds an array of strings.

Gtk4.GtkStringObject — Type

See the GTK docs.

Gtk4.GtkStringSorter — Type

See the GTK docs.

Gtk4.GtkStyleContext — Type

See the GTK docs.

Gtk4.GtkSwitch — Type

See the GTK docs.

Gtk4.GtkText — Type

See the GTK docs.

Gtk4.GtkTextBuffer — Type

See the GTK docs.

Gtk4.GtkTextChildAnchor — Type

See the GTK docs.

Gtk4.GtkTextIter — Type

See the GTK docs.

Gtk4.GtkTextIter — Method

GtkTextIter(text::GtkTextBuffer, char_offset::Integer)

Creates a GtkTextIter with offset char_offset (one-based index).

Gtk4.GtkTextMark — Type

See the GTK docs.

Gtk4.GtkTextTag — Type

See the GTK docs.

Gtk4.GtkTextTagTable — Type

See the GTK docs.

Gtk4.GtkTextView — Type

See the GTK docs.

Gtk4.GtkToggleButton — Type

See the GTK docs.

Gtk4.GtkTooltip — Type

See the GTK docs.

Gtk4.GtkTreeExpander — Type

See the GTK docs.

Gtk4.GtkTreeIter — Type

See the GTK docs.

Gtk4.GtkTreeListModel — Type

See the GTK docs.

Gtk4.GtkTreeListRow — Type

See the GTK docs.

Gtk4.GtkTreeListRowSorter — Type

See the GTK docs.

Gtk4.GtkTreeModelFilter — Type

See the GTK docs.

Gtk4.GtkTreeModelSort — Type

See the GTK docs.

Gtk4.GtkTreePath — Type

See the GTK docs.

Gtk4.GtkTreeSelection — Type

See the GTK docs.

Gtk4.GtkTreeStore — Type

See the GTK docs.

Gtk4.GtkTreeView — Type

See the GTK docs.

Gtk4.GtkTreeViewColumn — Type

See the GTK docs.

Gtk4.GtkUriLauncher — Type

See the GTK docs.

Gtk4.GtkVideo — Type

See the GTK docs.

Gtk4.GtkViewport — Type

See the GTK docs.

Gtk4.GtkVolumeButton — Type

See the GTK docs.

Gtk4.GtkWidget — Type

See the GTK docs.

Gtk4.GtkWidgetPaintable — Type

See the GTK docs.

Gtk4.GtkWindowControls — Type

See the GTK docs.

Gtk4.GtkWindowGroup — Type

See the GTK docs.

Gtk4.GtkWindowHandle — Type

See the GTK docs.

Gtk4.IconLookupFlags — Type

See the GTK docs.

Gtk4.IconSize — Type

See the GTK docs.

Gtk4.IconThemeError — Type

See the GTK docs.

Gtk4.IconViewDropPosition — Type

See the GTK docs.

Gtk4.ImageType — Type

See the GTK docs.

Gtk4.InputHints — Type

See the GTK docs.

Gtk4.InputPurpose — Type

See the GTK docs.

Gtk4.InputSource — Type

See the GTK docs.

Gtk4.InscriptionOverflow — Type

See the GTK docs.

Gtk4.Justification — Type

See the GTK docs.

Gtk4.KeyMatch — Type

See the GTK docs.

Gtk4.LevelBarMode — Type

See the GTK docs.

Gtk4.License — Type

See the GTK docs.

Gtk4.ListScrollFlags — Type

See the GTK docs.

Gtk4.ListTabBehavior — Type

See the GTK docs.

Gtk4.MemoryFormat — Type

See the GTK docs.

Gtk4.MessageType — Type

See the GTK docs.

Gtk4.ModifierType — Type

See the GTK docs.

Gtk4.MovementStep — Type

See the GTK docs.

Gtk4.NaturalWrapMode — Type

See the GTK docs.

Gtk4.NotebookTab — Type

See the GTK docs.

Gtk4.NotifyType — Type

See the GTK docs.

Gtk4.NumberUpLayout — Type

See the GTK docs.

Gtk4.Ordering — Type

See the GTK docs.

Gtk4.Orientation — Type

See the GTK docs.

Gtk4.Overflow — Type

See the GTK docs.

Gtk4.PackType — Type

See the GTK docs.

Gtk4.PadActionType — Type

See the GTK docs.

Gtk4.PageOrientation — Type

See the GTK docs.

Gtk4.PageSet — Type

See the GTK docs.

Gtk4.PaintableFlags — Type

See the GTK docs.

Gtk4.PanDirection — Type

See the GTK docs.

Gtk4.PickFlags — Type

See the GTK docs.

Gtk4.PolicyType — Type

See the GTK docs.

Gtk4.PopoverMenuFlags — Type

See the GTK docs.

Gtk4.PositionType — Type

See the GTK docs.

Gtk4.PrintCapabilities — Type

See the GTK docs.

Gtk4.PrintDuplex — Type

See the GTK docs.

Gtk4.PrintError — Type

See the GTK docs.

Gtk4.PrintOperationAction — Type

See the GTK docs.

Gtk4.PrintOperationResult — Type

See the GTK docs.

Gtk4.PrintPages — Type

See the GTK docs.

Gtk4.PrintQuality — Type

See the GTK docs.

Gtk4.PrintStatus — Type

See the GTK docs.

Gtk4.PropagationLimit — Type

See the GTK docs.

Gtk4.PropagationPhase — Type

See the GTK docs.

Gtk4.RecentManagerError — Type

See the GTK docs.

Gtk4.ResponseType — Type

See the GTK docs.

Gtk4.RevealerTransitionType — Type

See the GTK docs.

Gtk4.ScrollDirection — Type

See the GTK docs.

Gtk4.ScrollStep — Type

See the GTK docs.

Gtk4.ScrollType — Type

See the GTK docs.

Gtk4.ScrollUnit — Type

See the GTK docs.

Gtk4.ScrollablePolicy — Type

See the GTK docs.

Gtk4.SeatCapabilities — Type

See the GTK docs.

Gtk4.SelectionMode — Type

See the GTK docs.

Gtk4.SensitivityType — Type

See the GTK docs.

Gtk4.ShortcutActionFlags — Type

See the GTK docs.

Gtk4.ShortcutScope — Type

See the GTK docs.

Gtk4.ShortcutType — Type

See the GTK docs.

Gtk4.SizeGroupMode — Type

See the GTK docs.

Gtk4.SizeRequestMode — Type

See the GTK docs.

Gtk4.SortType — Type

See the GTK docs.

Gtk4.SorterChange — Type

See the GTK docs.

Gtk4.SorterOrder — Type

See the GTK docs.

Gtk4.SpinButtonUpdatePolicy — Type

See the GTK docs.

Gtk4.SpinType — Type

See the GTK docs.

Gtk4.StackTransitionType — Type

See the GTK docs.

Gtk4.StateFlags — Type

See the GTK docs.

Gtk4.StringFilterMatchMode — Type

See the GTK docs.

Gtk4.StyleContextPrintFlags — Type

See the GTK docs.

Gtk4.SubpixelLayout — Type

See the GTK docs.

Gtk4.SurfaceEdge — Type

See the GTK docs.

Gtk4.SymbolicColor — Type

See the GTK docs.

Gtk4.SystemSetting — Type

See the GTK docs.

Gtk4.TextDirection — Type

See the GTK docs.

Gtk4.TextExtendSelection — Type

See the GTK docs.

Gtk4.TextSearchFlags — Type

See the GTK docs.

Gtk4.TextViewLayer — Type

See the GTK docs.

Gtk4.TextWindowType — Type

See the GTK docs.

Gtk4.TextureError — Type

See the GTK docs.

Gtk4.ToplevelState — Type

See the GTK docs.

Gtk4.TouchpadGesturePhase — Type

See the GTK docs.

Gtk4.TreeViewColumnSizing — Type

See the GTK docs.

Gtk4.TreeViewDropPosition — Type

See the GTK docs.

Gtk4.TreeViewGridLines — Type

See the GTK docs.

Gtk4.Unit — Type

See the GTK docs.

Gtk4.VulkanError — Type

See the GTK docs.

Gtk4.WrapMode — Type

See the GTK docs.

Gtk4.GdkPixbufLib.GdkPixbuf — Type

See the GTK docs.

Gtk4.GdkPixbufLib.GdkPixbufAnimation — Type

See the GTK docs.

Gtk4.GdkPixbufLib.GdkPixbufAnimationIter — Type

See the GTK docs.

Gtk4.GdkPixbufLib.GdkPixbufFormat — Type

See the GTK docs.

Gtk4.GdkPixbufLib.GdkPixbufLoader — Type

See the GTK docs.

Gtk4.GdkPixbufLib.GdkPixbufSimpleAnim — Type

See the GTK docs.