API Reference

Public interface

Gtk4._GtkTextIter — Method

_GtkTextIter(text::GtkTextBuffer, char_offset::Integer)

Creates a _GtkTextIter with offset char_offset (one-based index).

Base.parent — Function

parent(w::GtkWidget)

Returns the parent widget of w, or nothing if the widget has not been set as the child of another widget (or is a toplevel widget, like a GtkWindow).

See also toplevel.

Related GTK function: gtk_widget_get_parent()

Base.show — Method

show(w::GtkWidget)

Flag w to be displayed and return w.

Related GTK function: gtk_widget_show()

Cairo.destroy — Method

destroy(win::GtkWindow)

Destroy win and unreference all widgets in its hierarchy.

Related GTK function: gtk_window_destroy()

Graphics.height — Function

height(w::GtkWidget)

Returns the allocated height of w in pixels.

Related GTK function: gtk_widget_get_allocated_height()

Graphics.width — Function

width(w::GtkWidget)

Returns the allocated width of w in pixels.

Related GTK function: gtk_widget_get_allocated_width()

Gtk4.buffer — Method

buffer(iter::Union{Ref{_GtkTextIter}, _GtkTextIter})

Returns the buffer associated with iter.

Gtk4.create_mark — Method

create_mark(buffer::GtkTextBuffer, mark_name, it::TI, left_gravity::Bool)
create_mark(buffer::GtkTextBuffer, it::TI)

Impements gtk_text_buffer_create_mark.

Gtk4.cursor — Function

cursor(w::GtkWidget)

Gets the cursor c for a widget w.

Related GTK function: gtk_widget_get_cursor()

Gtk4.cursor — Method

cursor(w::GtkWidget, c)

Sets a cursor c when the mouse is over a widget w. If c is nothing, use the default cursor for w.

Related GTK function: gtk_widget_set_cursor()

Gtk4.fullscreen — Method

fullscreen(win::GtkWindow)

Set win to fullscreen mode.

See also unfullscreen.

Related GTK function: gtk_window_fullscreen()

Gtk4.grab_focus — Method

grab_focus(w::GtkWidget)

Gives w the keyboard focus for the window it is in. Returns false if this fails.

Related GTK function: gtk_widget_grab_focus()

Gtk4.hasparent — Method

hasparent(w::GtkWidget) -> Bool

Returns true if w has a parent widget.

See also parent.

Gtk4.hide — Method

hide(w::GtkWidget)

Flag w to be hidden and return w. This is the opposite of show.

Related GTK function: gtk_widget_hide()

Gtk4.isvisible — Method

isvisible(w::GtkWidget) -> Bool

Returns whether w and all of its parents are marked as visible.

Related GTK function: gtk_widget_is_visible()

Gtk4.maximize — Method

maximize(win::GtkWindow)

Request that the window win be maximized.

See also unmaximize.

Related GTK function: gtk_window_maximize()

Gtk4.place_cursor — Method

place_cursor(buffer::GtkTextBuffer, it::_GtkTextIter)
place_cursor(buffer::GtkTextBuffer, pos::Int)

Place the cursor at indicated position.

Gtk4.screen_size — Function

screen_size(widget=nothing)

Returns a tuple (width,height) that gives the primary monitor size for the display where widget is being displayed, or the default display if widget is unrealized or not given.

Gtk4.scroll_to — Method

scroll_to(view::GtkTextView, mark::GtkTextMark, within_margin::Real,
               use_align::Bool, xalign::Real, yalign::Real)

scroll_to(view::GtkTextView, iter::TI, within_margin::Real,
          use_align::Bool, xalign::Real, yalign::Real)

Implements gtk_text_view_scroll_to_mark and gtk_text_view_scroll_to_iter.

Gtk4.search — Function

search(buffer::GtkTextBuffer, str::AbstractString, direction = :forward,
    flag = GtkTextSearchFlags.GTK_TEXT_SEARCH_TEXT_ONLY)

Search text str in buffer in direction :forward or :backward starting from the cursor position in the buffer.

Returns a tuple (found, start, stop) where found indicates whether the search was successful and start and stop are _GtkTextIters containing the location of the match.

Gtk4.select_range — Method

select_range(buffer::GtkTextBuffer, ins::TI, bound::TI)
select_range(buffer::GtkTextBuffer, range::GtkTextRange)

Select the text in buffer accorind to _GtkTextIter ins and bound.

Implements gtk_text_buffer_select_range.

Gtk4.selection_bounds — Method

selection_bounds(buffer::GtkTextBuffer)

Returns a tuple (selected, start, stop) indicating if text is selected in the buffer, and if so sets the _GtkTextIter start and stop to point to the selected text.

Implements gtk_text_buffer_get_selection_bounds.

Gtk4.toplevel — Method

toplevel(w::GtkWidget)

Returns the topmost ancestor of w, which in most cases will be a GtkWindow.

See also parent.

Related GTK function: gtk_widget_get_root()

Gtk4.unfullscreen — Method

unfullscreen(win::GtkWindow)

If win is in fullscreen mode, return it to normal mode.

See also fullscreen.

Related GTK function: gtk_window_unfullscreen()

Gtk4.unmaximize — Method

unmaximize(win::GtkWindow)

If win is maximized, return it to its former size.

See also maximize.

Related GTK function: gtk_window_unmaximize()

Gtk4.visible — Method

visible(w::GtkWidget, state::Bool)

Control visibility of w. Note that w will not be visible unless its parent is also visible.

Related GTK function: gtk_widget_set_visible()

Gtk4.widget — Function

widget(c::GtkEventController)

Returns the widget associated with an event controller.

Gtk4.@load_builder — Macro

@load_builder(b::GtkBuilder)

Loads all GtkBuildable objects from a GtkBuilder object and assigns them to Julia variables in the current scope. GtkBuilder ID's are mapped onto Julia variable names.

Private methods

Base.skip — Method

skip(iter::Ref{_GtkTextIter}, count::Integer, what::Symbol)

Moves iter according to the operation specified by what and count. Operations are :

:chars (gtk_text_iter_forward_chars)

:lines (gtk_text_iter_forward_lines)

:words (gtk_text_iter_forward_word_ends)

:word_cursor_positions (gtk_text_iter_forward_cursor_positions)

:sentences (gtk_text_iter_forward_sentence_ends)

:visible_words (gtk_text_iter_forward_visible_word_ends)

:visible_cursor_positions (gtk_text_iter_forward_visible_cursor_positions)

:visible_lines (gtk_text_iter_forward_visible_lines)

:line_ends (gtk_text_iter_forward_visible_lines)

Base.skip — Method

skip(iter::Ref{_GtkTextIter}, count::Integer)

Moves iter count characters. Returns a Bool indicating if the move was successful.

Base.skip — Method

skip(iter::Ref{_GtkTextIter}, what::Symbol)

Moves iter according to the operation specified by what. Operations are :

:forward_line (gtk_text_iter_forward_line)

:backward_line (gtk_text_iter_backward_line)

:forward_to_line_end (gtk_text_iter_forward_to_line_end)

:backward_word_start (gtk_text_iter_forward_word_end)

:forward_word_end (gtk_text_iter_backward_word_start)

:backward_sentence_start (gtk_text_iter_backward_sentence_start)

:forward_sentence_end (gtk_text_iter_forward_sentence_end)

Gtk4.backward_search — Method

backward_search(iter::Ref{_GtkTextIter},
    str::AbstractString, start::Ref{_GtkTextIter},
    stop::Ref{_GtkTextIter}, limit::Ref{_GtkTextIter}, flag::Int32)

Implements gtk_text_iter_backward_search.

Gtk4.buffer_to_window_coords — Function

buffer_to_window_coords(view::GtkTextView, buffer_x::Integer, buffer_y::Integer, wintype::Integer = 0)

Implements gtk_text_view_buffer_to_window_coords.

Gtk4.char_offset — Method

char_offset(iter::Union{Ref{_GtkTextIter}, _GtkTextIter})

Returns the offset of iter (one-based index).

Gtk4.display — Function

display(w::GtkWidget)

Gets the GdkDisplay for w. Should only be called if w has been added to a widget hierarchy.

Related GTK function: gtk_widget_get_display()

Gtk4.find_controller — Method

find_controller(w::GtkWidget, ::Type{T}) where T <: GtkEventController

Returns an event controller of type T connected to a widget, or nothing if one doesn't exist. This function is intended for testing purposes (to simulate events) and is not recommended otherwise, as there is a performance penalty for creating a list of a widget's event controllers.

Related GTK function: gtk_widget_observe_controllers)

Gtk4.forward_search — Method

forward_search(iter::Ref{_GtkTextIter},
    str::AbstractString, start::Ref{_GtkTextIter},
    stop::Ref{_GtkTextIter}, limit::Ref{_GtkTextIter}, flag::Int32)

Implements gtk_text_iter_forward_search.

Gtk4.monitor — Method

monitor(w::GtkWidget)

Gets the GdkMonitor where w is displayed, or nothing if the widget is not part of a widget hierarchy.

Gtk4.text_iter_at_position — Method

text_iter_at_position(view::GtkTextView, x::Integer, y::Integer)

Implements gtk_text_view_get_iter_at_position.

Gtk4.window_to_buffer_coords — Function

window_to_buffer_coords(view::GtkTextView, window_x::Integer, window_y::Integer, wintype::Integer = 2)

Implements gtk_text_view_window_to_buffer_coords.